Change package from jpa3 to hibernate6.jpa
