+++ data-hibernate6-jpa/src/main/java/io/micronaut/data/hibernate6/jpa/annotation/EntityGraph.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.data.jpa3.annotation;
+package io.micronaut.data.hibernate6.jpa.annotation;
 
 import io.micronaut.context.annotation.AliasFor;
