+++ data-hibernate6-jpa/src/main/java/io/micronaut/data/hibernate6/jpa/repository/criteria/SpecificationComposition.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.data.jpa3.repository.criteria;
+package io.micronaut.data.hibernate6.jpa.repository.criteria;
 
 import io.micronaut.core.annotation.NonNull;
 import io.micronaut.core.annotation.Nullable;
