+++ data-hibernate6-jpa/src/main/java/io/micronaut/data/hibernate6/jpa/repository/intercept/CountSpecificationInterceptor.java
@@ -13,36 +13,34 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.data.jpa3.repository.intercept;
+package io.micronaut.data.hibernate6.jpa.repository.intercept;
 
 import io.micronaut.aop.MethodInvocationContext;
 import io.micronaut.core.annotation.Internal;
 import io.micronaut.core.annotation.NonNull;
 import io.micronaut.core.type.ReturnType;
-import io.micronaut.data.exceptions.EmptyResultException;
 import io.micronaut.data.intercept.RepositoryMethodKey;
-import io.micronaut.data.jpa3.operations.JpaRepositoryOperations;
-import io.micronaut.data.jpa3.repository.criteria.Specification;
+import io.micronaut.data.hibernate6.jpa.operations.JpaRepositoryOperations;
+import io.micronaut.data.hibernate6.jpa.repository.criteria.Specification;
 import io.micronaut.data.operations.RepositoryOperations;
 
 import jakarta.persistence.EntityManager;
-import jakarta.persistence.NoResultException;
 import jakarta.persistence.TypedQuery;
 import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.persistence.criteria.CriteriaQuery;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
-import java.util.Optional;
+import java.util.Collections;
 
 /**
- * Implementation of {@code findOne(Specification)} for JPA specifications.
+ * Interceptor that supports count specifications.
  *
  * @author graemerocher
  * @author Denis Stepanov
  * @since 3.1
  */
 @Internal
-public class FindOneSpecificationInterceptor extends AbstractSpecificationInterceptor<Object, Object> {
+public class CountSpecificationInterceptor extends AbstractSpecificationInterceptor<Object, Number> {
     private final JpaRepositoryOperations jpaOperations;
 
     /**
@@ -50,7 +48,7 @@ public class FindOneSpecificationInterceptor extends AbstractSpecificationInterc
      *
      * @param operations The operations
      */
-    protected FindOneSpecificationInterceptor(@NonNull RepositoryOperations operations) {
+    public CountSpecificationInterceptor(@NonNull RepositoryOperations operations) {
         super(operations);
         if (operations instanceof JpaRepositoryOperations) {
             this.jpaOperations = (JpaRepositoryOperations) operations;
@@ -60,39 +58,34 @@ public class FindOneSpecificationInterceptor extends AbstractSpecificationInterc
     }
 
     @Override
-    public Object intercept(RepositoryMethodKey methodKey, MethodInvocationContext<Object, Object> context) {
+    public Number intercept(RepositoryMethodKey methodKey, MethodInvocationContext<Object, Number> context) {
         Specification specification = getSpecification(context);
         final EntityManager entityManager = jpaOperations.getCurrentEntityManager();
         final CriteriaBuilder criteriaBuilder = entityManager.getCriteriaBuilder();
-        final CriteriaQuery<Object> query = criteriaBuilder.createQuery((Class<Object>) getRequiredRootEntity(context));
-        final Root<Object> root = query.from((Class<Object>) getRequiredRootEntity(context));
+        final CriteriaQuery<Long> query = criteriaBuilder.createQuery(Long.class);
+        final Root<?> root = query.from(getRequiredRootEntity(context));
         final Predicate predicate = specification.toPredicate(root, query, criteriaBuilder);
         if (predicate != null) {
             query.where(predicate);
         }
-        query.select(root);
+        if (query.isDistinct()) {
+            query.select(criteriaBuilder.countDistinct(root));
+        } else {
+            query.select(criteriaBuilder.count(root));
+        }
+        query.orderBy(Collections.emptyList());
 
-        final TypedQuery<?> typedQuery = entityManager.createQuery(query);
-        try {
-            final Object result = typedQuery.getSingleResult();
-            final ReturnType<?> rt = context.getReturnType();
-            final Class<?> returnType = rt.getType();
-            if (returnType.isInstance(result)) {
-                return result;
-            } else {
-                return operations.getConversionService().convertRequired(
-                        result,
-                        rt.asArgument()
-                );
-            }
-        } catch (NoResultException e) {
-            if (context.isNullable()) {
-                return null;
-            } else if (context.getReturnType().isOptional()) {
-                return Optional.empty();
-            } else {
-                throw new EmptyResultException();
-            }
+        final TypedQuery<Long> typedQuery = entityManager.createQuery(query);
+        final Long result = typedQuery.getSingleResult();
+        final ReturnType<Number> rt = context.getReturnType();
+        final Class<Number> returnType = rt.getType();
+        if (returnType.isInstance(result)) {
+            return result;
+        } else {
+            return operations.getConversionService().convertRequired(
+                    result,
+                    rt.asArgument()
+            );
         }
     }
 }
