+++ data-hibernate6-jpa/src/main/java/io/micronaut/data/hibernate6/jpa/repository/intercept/FindAllSpecificationInterceptor.java
@@ -13,34 +13,32 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.data.jpa3.repository.intercept;
+package io.micronaut.data.hibernate6.jpa.repository.intercept;
 
 import io.micronaut.aop.MethodInvocationContext;
 import io.micronaut.core.annotation.Internal;
 import io.micronaut.core.annotation.NonNull;
-import io.micronaut.core.type.ReturnType;
 import io.micronaut.data.intercept.RepositoryMethodKey;
-import io.micronaut.data.jpa3.operations.JpaRepositoryOperations;
-import io.micronaut.data.jpa3.repository.criteria.Specification;
+import io.micronaut.data.hibernate6.jpa.operations.JpaRepositoryOperations;
+import io.micronaut.data.hibernate6.jpa.repository.criteria.Specification;
+import io.micronaut.data.model.Sort;
 import io.micronaut.data.operations.RepositoryOperations;
 
 import jakarta.persistence.EntityManager;
-import jakarta.persistence.TypedQuery;
 import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.persistence.criteria.CriteriaQuery;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
-import java.util.Collections;
 
 /**
- * Interceptor that supports count specifications.
+ * Implementation of the unpaged version of {@code findAll(Specification)}.
  *
  * @author graemerocher
  * @author Denis Stepanov
  * @since 3.1
  */
 @Internal
-public class CountSpecificationInterceptor extends AbstractSpecificationInterceptor<Object, Number> {
+public class FindAllSpecificationInterceptor extends AbstractSpecificationInterceptor<Object, Object> {
     private final JpaRepositoryOperations jpaOperations;
 
     /**
@@ -48,7 +46,7 @@ public class CountSpecificationInterceptor extends AbstractSpecificationIntercep
      *
      * @param operations The operations
      */
-    public CountSpecificationInterceptor(@NonNull RepositoryOperations operations) {
+    protected FindAllSpecificationInterceptor(@NonNull RepositoryOperations operations) {
         super(operations);
         if (operations instanceof JpaRepositoryOperations) {
             this.jpaOperations = (JpaRepositoryOperations) operations;
@@ -58,34 +56,39 @@ public class CountSpecificationInterceptor extends AbstractSpecificationIntercep
     }
 
     @Override
-    public Number intercept(RepositoryMethodKey methodKey, MethodInvocationContext<Object, Number> context) {
-        Specification specification = getSpecification(context);
+    public Object intercept(RepositoryMethodKey methodKey, MethodInvocationContext<Object, Object> context) {
+        final Specification specification = getSpecification(context);
         final EntityManager entityManager = jpaOperations.getCurrentEntityManager();
         final CriteriaBuilder criteriaBuilder = entityManager.getCriteriaBuilder();
-        final CriteriaQuery<Long> query = criteriaBuilder.createQuery(Long.class);
-        final Root<?> root = query.from(getRequiredRootEntity(context));
+        final CriteriaQuery<Object> query = criteriaBuilder.createQuery((Class<Object>) getRequiredRootEntity(context));
+        final Root<Object> root = query.from((Class<Object>) getRequiredRootEntity(context));
         final Predicate predicate = specification.toPredicate(root, query, criteriaBuilder);
         if (predicate != null) {
             query.where(predicate);
         }
-        if (query.isDistinct()) {
-            query.select(criteriaBuilder.countDistinct(root));
-        } else {
-            query.select(criteriaBuilder.count(root));
-        }
-        query.orderBy(Collections.emptyList());
+        query.select(root);
 
-        final TypedQuery<Long> typedQuery = entityManager.createQuery(query);
-        final Long result = typedQuery.getSingleResult();
-        final ReturnType<Number> rt = context.getReturnType();
-        final Class<Number> returnType = rt.getType();
-        if (returnType.isInstance(result)) {
-            return result;
-        } else {
-            return operations.getConversionService().convertRequired(
-                    result,
-                    rt.asArgument()
-            );
+        if (context.getParameterValues().length > 1) {
+            addSort(context.getParameterValues()[1], query, root, criteriaBuilder);
+        }
+        return entityManager.createQuery(query).getResultList();
+    }
+
+    /**
+     * Add sort to the query.
+     *
+     * @param sortObject      The sort object
+     * @param query           The query
+     * @param root            The root
+     * @param criteriaBuilder The criteria builder
+     */
+    protected void addSort(Object sortObject,
+                           CriteriaQuery<Object> query, Root<Object> root, CriteriaBuilder criteriaBuilder) {
+        if (sortObject instanceof Sort) {
+            Sort sort = (Sort) sortObject;
+            if (sort.isSorted()) {
+                query.orderBy(getOrders(sort, root, criteriaBuilder));
+            }
         }
     }
 }
