+++ data-hibernate6-jpa/src/main/java/io/micronaut/data/hibernate6/jpa/repository/intercept/FindOneSpecificationInterceptor.java
@@ -13,32 +13,36 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.data.jpa3.repository.intercept;
+package io.micronaut.data.hibernate6.jpa.repository.intercept;
 
 import io.micronaut.aop.MethodInvocationContext;
 import io.micronaut.core.annotation.Internal;
 import io.micronaut.core.annotation.NonNull;
+import io.micronaut.core.type.ReturnType;
+import io.micronaut.data.exceptions.EmptyResultException;
 import io.micronaut.data.intercept.RepositoryMethodKey;
-import io.micronaut.data.jpa3.operations.JpaRepositoryOperations;
-import io.micronaut.data.jpa3.repository.criteria.Specification;
-import io.micronaut.data.model.Sort;
+import io.micronaut.data.hibernate6.jpa.operations.JpaRepositoryOperations;
+import io.micronaut.data.hibernate6.jpa.repository.criteria.Specification;
 import io.micronaut.data.operations.RepositoryOperations;
 
 import jakarta.persistence.EntityManager;
+import jakarta.persistence.NoResultException;
+import jakarta.persistence.TypedQuery;
 import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.persistence.criteria.CriteriaQuery;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
+import java.util.Optional;
 
 /**
- * Implementation of the unpaged version of {@code findAll(Specification)}.
+ * Implementation of {@code findOne(Specification)} for JPA specifications.
  *
  * @author graemerocher
  * @author Denis Stepanov
  * @since 3.1
  */
 @Internal
-public class FindAllSpecificationInterceptor extends AbstractSpecificationInterceptor<Object, Object> {
+public class FindOneSpecificationInterceptor extends AbstractSpecificationInterceptor<Object, Object> {
     private final JpaRepositoryOperations jpaOperations;
 
     /**
@@ -46,7 +50,7 @@ public class FindAllSpecificationInterceptor extends AbstractSpecificationInterc
      *
      * @param operations The operations
      */
-    protected FindAllSpecificationInterceptor(@NonNull RepositoryOperations operations) {
+    protected FindOneSpecificationInterceptor(@NonNull RepositoryOperations operations) {
         super(operations);
         if (operations instanceof JpaRepositoryOperations) {
             this.jpaOperations = (JpaRepositoryOperations) operations;
@@ -57,7 +61,7 @@ public class FindAllSpecificationInterceptor extends AbstractSpecificationInterc
 
     @Override
     public Object intercept(RepositoryMethodKey methodKey, MethodInvocationContext<Object, Object> context) {
-        final Specification specification = getSpecification(context);
+        Specification specification = getSpecification(context);
         final EntityManager entityManager = jpaOperations.getCurrentEntityManager();
         final CriteriaBuilder criteriaBuilder = entityManager.getCriteriaBuilder();
         final CriteriaQuery<Object> query = criteriaBuilder.createQuery((Class<Object>) getRequiredRootEntity(context));
@@ -68,26 +72,26 @@ public class FindAllSpecificationInterceptor extends AbstractSpecificationInterc
         }
         query.select(root);
 
-        if (context.getParameterValues().length > 1) {
-            addSort(context.getParameterValues()[1], query, root, criteriaBuilder);
-        }
-        return entityManager.createQuery(query).getResultList();
-    }
-
-    /**
-     * Add sort to the query.
-     *
-     * @param sortObject      The sort object
-     * @param query           The query
-     * @param root            The root
-     * @param criteriaBuilder The criteria builder
-     */
-    protected void addSort(Object sortObject,
-                           CriteriaQuery<Object> query, Root<Object> root, CriteriaBuilder criteriaBuilder) {
-        if (sortObject instanceof Sort) {
-            Sort sort = (Sort) sortObject;
-            if (sort.isSorted()) {
-                query.orderBy(getOrders(sort, root, criteriaBuilder));
+        final TypedQuery<?> typedQuery = entityManager.createQuery(query);
+        try {
+            final Object result = typedQuery.getSingleResult();
+            final ReturnType<?> rt = context.getReturnType();
+            final Class<?> returnType = rt.getType();
+            if (returnType.isInstance(result)) {
+                return result;
+            } else {
+                return operations.getConversionService().convertRequired(
+                        result,
+                        rt.asArgument()
+                );
+            }
+        } catch (NoResultException e) {
+            if (context.isNullable()) {
+                return null;
+            } else if (context.getReturnType().isOptional()) {
+                return Optional.empty();
+            } else {
+                throw new EmptyResultException();
             }
         }
     }
