+++ data-hibernate6-jpa/src/main/java/io/micronaut/data/hibernate6/jpa/repository/intercept/FindPageSpecificationInterceptor.java
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.data.jpa3.repository.intercept;
+package io.micronaut.data.hibernate6.jpa.repository.intercept;
 
 import io.micronaut.aop.MethodInvocationContext;
 import io.micronaut.core.annotation.Internal;
 import io.micronaut.core.annotation.NonNull;
 import io.micronaut.data.intercept.RepositoryMethodKey;
-import io.micronaut.data.jpa3.operations.JpaRepositoryOperations;
-import io.micronaut.data.jpa3.repository.criteria.Specification;
+import io.micronaut.data.hibernate6.jpa.operations.JpaRepositoryOperations;
+import io.micronaut.data.hibernate6.jpa.repository.criteria.Specification;
 import io.micronaut.data.model.Page;
 import io.micronaut.data.model.Pageable;
 import io.micronaut.data.model.Sort;
