+++ data-hibernate6-jpa/src/main/java/io/micronaut/data/hibernate6/operations/HibernateJpaOperations.java
@@ -26,8 +26,8 @@ import io.micronaut.core.convert.ConversionService;
 import io.micronaut.core.type.Argument;
 import io.micronaut.core.util.ArgumentUtils;
 import io.micronaut.data.annotation.QueryHint;
-import io.micronaut.data.jpa3.annotation.EntityGraph;
-import io.micronaut.data.jpa3.operations.JpaRepositoryOperations;
+import io.micronaut.data.hibernate6.jpa.annotation.EntityGraph;
+import io.micronaut.data.hibernate6.jpa.operations.JpaRepositoryOperations;
 import io.micronaut.data.model.Page;
 import io.micronaut.data.model.Pageable;
 import io.micronaut.data.model.runtime.DeleteBatchOperation;
