+++ data-hibernate6-jpa/src/test/java/io/micronaut/data/hibernate6/BookRepository.java
@@ -21,8 +21,8 @@ import io.micronaut.data.annotation.Join;
 import io.micronaut.data.annotation.Query;
 import io.micronaut.data.annotation.Repository;
 import io.micronaut.data.annotation.Where;
-import io.micronaut.data.jpa3.annotation.EntityGraph;
-import io.micronaut.data.jpa3.repository.JpaSpecificationExecutor;
+import io.micronaut.data.hibernate6.jpa.annotation.EntityGraph;
+import io.micronaut.data.hibernate6.jpa.repository.JpaSpecificationExecutor;
 import io.micronaut.data.model.Page;
 import io.micronaut.data.model.Pageable;
 import io.micronaut.data.tck.entities.Author;
