+++ data-hibernate6-jpa/src/test/java/io/micronaut/data/hibernate6/ChildrenRepository.java
@@ -4,7 +4,7 @@ import io.micronaut.data.annotation.Query;
 import io.micronaut.data.annotation.Repository;
 import io.micronaut.data.hibernate6.entities.Children;
 import io.micronaut.data.hibernate6.entities.ChildrenId;
-import io.micronaut.data.jpa3.repository.JpaRepository;
+import io.micronaut.data.hibernate6.jpa.repository.JpaRepository;
 
 @Repository
 public interface ChildrenRepository extends JpaRepository<Children, ChildrenId> {
