+++ data-hibernate6-jpa/src/test/java/io/micronaut/data/hibernate6/PersonCrudRepository.java
@@ -20,7 +20,7 @@ import io.micronaut.data.annotation.Id;
 import io.micronaut.data.annotation.Query;
 import io.micronaut.data.annotation.QueryHint;
 import io.micronaut.data.annotation.Repository;
-import io.micronaut.data.jpa3.repository.JpaRepository;
+import io.micronaut.data.hibernate6.jpa.repository.JpaRepository;
 import io.micronaut.data.model.Page;
 import io.micronaut.data.model.Pageable;
 import io.micronaut.data.tck.entities.Person;
