+++ data-processor/src/main/java/io/micronaut/data/processor/visitors/finders/AbstractSpecificationMethodMatcher.java
@@ -92,7 +92,7 @@ public abstract class AbstractSpecificationMethodMatcher extends AbstractPrefixP
 
     protected final boolean isFirstParameterMicronautJpaSpecification(@NonNull MethodElement methodElement) {
         return isFirstParameterAssignable(methodElement, "io.micronaut.data.jpa.repository.criteria.Specification")
-            || isFirstParameterAssignable(methodElement, "io.micronaut.data.jpa3.repository.criteria.Specification");
+            || isFirstParameterAssignable(methodElement, "io.micronaut.data.hibernate6.jpa.repository.criteria.Specification");
     }
 
     protected final boolean isFirstParameterMicronautDataQuerySpecification(@NonNull MethodElement methodElement) {
@@ -121,7 +121,7 @@ public abstract class AbstractSpecificationMethodMatcher extends AbstractPrefixP
 
     private boolean isFirstParameterMicronautDataPredicateSpecification(@NonNull MethodElement methodElement) {
         return isFirstParameterAssignable(methodElement, "io.micronaut.data.repository.jpa.criteria.PredicateSpecification")
-            || isFirstParameterAssignable(methodElement, "io.micronaut.data.repository.jpa3.criteria.PredicateSpecification");
+            || isFirstParameterAssignable(methodElement, "io.micronaut.data.hibernate6.jpa.criteria.PredicateSpecification");
     }
 
     private boolean isFirstParameterAssignable(@NonNull MethodElement methodElement, String clazz) {
