+++ data-processor/src/main/java/io/micronaut/data/processor/visitors/finders/spec/FindAllSpecificationMethodMatcher.java
@@ -69,7 +69,7 @@ public class FindAllSpecificationMethodMatcher extends AbstractSpecificationMeth
                 try {
                     classElement = getInterceptorElement(mc, "io.micronaut.data.jpa.repository.intercept.FindAllSpecificationInterceptor");
                 } catch (IllegalStateException e) {
-                    classElement = getInterceptorElement(mc, "io.micronaut.data.jpa3.repository.intercept.FindAllSpecificationInterceptor");
+                    classElement = getInterceptorElement(mc, "io.micronaut.data.hibernate6.jpa.repository.intercept.FindAllSpecificationInterceptor");
                 }
                 return new MethodMatchInfo(
                     DataMethod.OperationType.QUERY,
