+++ data-processor/src/main/java/io/micronaut/data/processor/visitors/finders/spec/FindOneSpecificationMethodMatcher.java
@@ -59,7 +59,7 @@ public class FindOneSpecificationMethodMatcher extends AbstractSpecificationMeth
             try {
                 classElement = getInterceptorElement(mc, "io.micronaut.data.jpa.repository.intercept.FindOneSpecificationInterceptor");
             } catch (IllegalStateException e) {
-                classElement = getInterceptorElement(mc, "io.micronaut.data.jpa3.repository.intercept.FindOneSpecificationInterceptor");
+                classElement = getInterceptorElement(mc, "io.micronaut.data.hibernate6.jpa.repository.intercept.FindOneSpecificationInterceptor");
             }
             return new MethodMatchInfo(
                 DataMethod.OperationType.QUERY,
