+++ doc-examples/hibernate6-example-java/src/main/java/example/ProductRepository.java
@@ -3,9 +3,9 @@ package example;
 
 import io.micronaut.data.annotation.Join;
 import io.micronaut.data.annotation.Repository;
-import io.micronaut.data.jpa3.annotation.EntityGraph;
-import io.micronaut.data.jpa3.repository.JpaSpecificationExecutor;
-import io.micronaut.data.jpa3.repository.criteria.Specification;
+import io.micronaut.data.hibernate6.jpa.annotation.EntityGraph;
+import io.micronaut.data.hibernate6.jpa.repository.JpaSpecificationExecutor;
+import io.micronaut.data.hibernate6.jpa.repository.criteria.Specification;
 import io.micronaut.data.repository.CrudRepository;
 import io.reactivex.Maybe;
 import io.reactivex.Single;
